+++ semcore/typography/src/index.d.ts
@@ -47,6 +47,8 @@ export type TextProps = BoxProps & {
   use?: 'primary' | 'secondary';
   /** Makes text semi-transparent to indicate disabled state */
   disabled?: boolean;
+  /** Enable formatting/styling for all nested HTML tags with our default styles for them */
+  formatTags?: boolean;
   /** The text will not be wrapped on a new line and will be cut off with ellipsis. Also, it will show a hint with full text. */
   ellipsis?: true | EllipsisSettings;
   /** Settings for a hint with full text (cropped by ellipsis) */