+++ semcore/data-table/CHANGELOG.md
@@ -16,15 +16,17 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 ### Fixed
 
+- Rendering table with selectedRows without data.
 - `z-index` of expanded accordions and spinner container.
 - Usage of `changeSortSize` property with multiple columns.
 - Accordion content expands with visual issues when clicking on a merged cell.
 - Incorrect type for `selectedRows/onSelectedRowsChange` when `uniqueRowKey` type is not a string.
 - Hover styles aren't applied to themed cells.
+- Types for `CellRenderProps` - added ability to compare props.columnName with `ACCORDION` symbol.
 
 ### Changed
 
-- `overflow` for `Body.Cell` to default.
+- `rowProps` and `renderCell` will not be called when rendering an empty table.
 
 ## [16.0.7] - 2025-07-23